adding everything up, working refactored script
